requirements/prod.txt
Django==4.2.5
django_cleanup
sorl-thumbnail
transliterate
Pillow
django-ckeditor
python-dotenv==0.21.1
